Use base64 std encoding instead of urlencoding
