+++ pkg/proxy/session/token.go
@@ -462,7 +462,7 @@ func CompressToken(token string, pool *utils.LimitedBufferPool) (string, error) 
 		return "", apperrors.ErrTokenParts
 	}
 
-	info, err := base64.RawURLEncoding.DecodeString(parts[1])
+	info, err := base64.RawStdEncoding.DecodeString(parts[1])
 	if err != nil {
 		return "", err
 	}
@@ -482,7 +482,7 @@ func CompressToken(token string, pool *utils.LimitedBufferPool) (string, error) 
 
 	gzWr.Close()
 
-	compInfo := base64.RawURLEncoding.EncodeToString(compBuffer.Bytes())
+	compInfo := base64.RawStdEncoding.EncodeToString(compBuffer.Bytes())
 
 	return fmt.Sprintf("%s.%s.%s", parts[0], compInfo, parts[2]), nil
 }
@@ -493,7 +493,7 @@ func DecompressToken(compressedToken string) (string, error) {
 		return "", apperrors.ErrTokenParts
 	}
 
-	info, err := base64.RawURLEncoding.DecodeString(parts[1])
+	info, err := base64.RawStdEncoding.DecodeString(parts[1])
 	if err != nil {
 		return "", err
 	}
@@ -514,7 +514,7 @@ func DecompressToken(compressedToken string) (string, error) {
 
 	gzR.Close()
 
-	compInfo := base64.RawURLEncoding.EncodeToString(result)
+	compInfo := base64.RawStdEncoding.EncodeToString(result)
 
 	return fmt.Sprintf("%s.%s.%s", parts[0], compInfo, parts[2]), nil
 }
@@ -525,7 +525,7 @@ func EncryptAndCompressToken(token string, encryptionKey string, pool *utils.Lim
 		return "", apperrors.ErrTokenParts
 	}
 
-	info, err := base64.RawURLEncoding.DecodeString(parts[1])
+	info, err := base64.RawStdEncoding.DecodeString(parts[1])
 	if err != nil {
 		return "", err
 	}
@@ -600,7 +600,7 @@ func DecryptAndDecompressToken(token string, encryptionKey string) (string, erro
 
 	gzR.Close()
 
-	info := base64.RawURLEncoding.EncodeToString(result)
+	info := base64.RawStdEncoding.EncodeToString(result)
 
 	return fmt.Sprintf("%s.%s.%s", compHeader, info, compSignature), nil
 }
